% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotme.R
\name{plotme}
\alias{plotme}
\title{plotme}
\usage{
plotme(plotdata, x, y, z)
}
\arguments{
\item{plotdata}{data}

\item{x}{column number of x coordinate}

\item{y}{column number of y coordinate}

\item{z}{column number of group category}
}
\value{
a plot
}
\description{
plotme
}
\examples{
plotme(data,1,2,5)

}
