% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotme2.R
\name{plotme2}
\alias{plotme2}
\title{plotme2}
\usage{
plotme2(plotdata, x, y)
}
\arguments{
\item{plotdata}{data}

\item{x}{column number of x coordinate}

\item{y}{column number of y coordinate}
}
\value{
a plot
}
\description{
plotme2
}
\examples{
plotme2(data,1,2)

}
